Move domutils into dom folder
