+++ dom/domutils.js
@@ -1,0 +1,6 @@
+import DomUtils from './MemoryDomUtils'
+
+const domutils = new DomUtils()
+domutils.DomUtils = DomUtils
+
+export default domutils
