+++ index.es.js
@@ -5,5 +5,3 @@ export * from './util/index'
 
 // patched cssSelect
 export { default as cssSelect } from './vendor/css-select'
-// patched DomUtils
-export { default as DomUtils } from './vendor/domutils'
